fix(#3984): route Object.defineProperties array-`length` to ArraySetLength
`maybeEmitVecLengthDefine` implements ES §10.4.2.4 ArraySetLength for WasmGC
vec receivers, but had exactly ONE call site — inside
`compileObjectDefineProperty`. The static object-literal expansion inside
`compileObjectDefineProperties` re-parses each descriptor inline instead of
delegating, so it never reached that machinery.

On the standalone target that made

    Object.defineProperties(arr, { length: { value: 2 } })

a SILENT NO-OP: the length simply stayed put. A wrong value, not a refusal —
so neither the root-cause classifier nor the standalone floor could observe it.
The singular `Object.defineProperty(arr, "length", ...)` form was already
correct, so this is a routing gap over working machinery, not new substrate.
It is therefore disjoint from the #3251 per-index overlay epic.

Architectural fix rather than a budget allowance: the ArraySetLength unit
(`maybeEmitVecLengthDefine` + its two private helpers) moves out of the
`object-ops.ts` god-file into `src/codegen/array-length-define.ts`, which now
owns both callers. object-ops.ts shrinks by 312 lines; the LOC and function
budget gates pass with no allowance granted.

Also removes a dead synthetic `Object.defineProperty(...)` node in the
`defineProperties` loop that was built, text-ranged, re-parented and never
read. Its comment claimed the loop directly called
`compileObjectDefineProperty` — it never did, and that false claim is how this
gap stayed invisible.

Measured (paired per-file A/B, both arms in one process, attribution by
kill-switch removal; instrument validated against standalone official
43,106/25,460 = 59.1% and goal scope 8,545/6,004 = 70.3%):

    +34 pass, 0 lost, 0 compile_timeout, in-sweep controls held
    34 flipped / 103 reachable gated = 33.0%


✓
